requirements.txt
torch
torchvision
numpy
scikit-learn
pillow
streamlit
